move layer_decay_optimizer_constructor
