+++ mmseg/core/__init__.py
@@ -1,6 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .evaluation import *  # noqa: F401, F403
-from .layer_decay_optimizer_constructor import \
-    LayerDecayOptimizerConstructor  # noqa: F401
+from .optimizers import *  # noqa: F401, F403
 from .seg import *  # noqa: F401, F403
 from .utils import *  # noqa: F401, F403
