+++ mmseg/core/optimizers/__init__.py
@@ -1,0 +1,7 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+from .layer_decay_optimizer_constructor import \
+    LearningRateDecayOptimizerConstructor, LayerDecayOptimizerConstructor
+
+__all__ = [
+    'LearningRateDecayOptimizerConstructor', 'LayerDecayOptimizerConstructor'
+]
