+++ mmseg/core/utils/__init__.py
@@ -1,10 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .dist_util import check_dist_init, sync_random_seed
-from .layer_decay_optimizer_constructor import \
-    LearningRateDecayOptimizerConstructor
 from .misc import add_prefix
 
-__all__ = [
-    'add_prefix', 'LearningRateDecayOptimizerConstructor', 'check_dist_init',
-    'sync_random_seed'
-]
+__all__ = ['add_prefix', 'check_dist_init', 'sync_random_seed']
